Prototype version

+++ GameOfLife.Application/Games/Create/CreateGameCommandValidator.cs
@@ -7,9 +7,9 @@ namespace GameOfLife.Application.Games.Create
         public CreateGameCommandValidator()
         {
             RuleFor(game => game.Width)
-                .GreaterThan(0);
+                .GreaterThan((ushort)0);
             RuleFor(game => game.Height)
-                .GreaterThan(0);
+                .GreaterThan((ushort)0);
             RuleFor(game => game.State)
                 .NotNull().NotEmpty();
         }
