+++ GameOfLife.Application/Games/Create/CreateGameRequest.cs
@@ -8,6 +8,6 @@
 /// than enough to store the population size and smaller than int
 /// </param>
 public sealed record CreateGameRequest(
-    int Width,
-    int Height,
+    ushort Width,
+    ushort Height,
     ushort[][] State);
